ROFO-128 마이페이지에 유저 정보 띄우기

+++ domain/src/main/java/com/weit2nd/domain/model/User.kt
@@ -7,5 +7,5 @@ data class User(
 data class UserInfo(
     val nickname: String,
     val profileImage: String?,
-    val coin: Long,
+    val coin: Int,
 )
